backend/requirements.txt
django
djangorestframework
mysql-connector-python
PyJWT
cryptography
pyotp
django-cors-headers
